+++ README.md
@@ -29,3 +29,9 @@ conda: `conda install -c anty-filidor network_diffusion`
 
 In case of questions for the maintainer please send mails here:
 <michal.czuba.1995@gmail.com>.
+
+# todo
+
+- cicd
+- package version
+- requirements separation
